+++ src/main/java/com/memorymakerpeople/memoryrollingpaper/paper/model/PaperIdMapping.java
@@ -1,7 +1,5 @@
 package com.memorymakerpeople.memoryrollingpaper.paper.model;
 
-import java.math.BigInteger;
-
 public interface PaperIdMapping {
-    BigInteger getPaperId();
+    Long getPaperId();
 }
